Move /firefox/browsers to /firefox (Fixes #14222)
