+++ bedrock/base/templates/includes/protocol/navigation/menus/firefox.html
@@ -7,7 +7,7 @@
 {% set utm_params = 'utm_source=www.mozilla.org&utm_medium=referral&utm_campaign=nav&utm_content=firefox' %}
 
 <li class="c-menu-category mzp-has-drop-down mzp-js-expandable">
-  <a class="c-menu-title" href="{{ url('firefox.browsers.index') }}" aria-haspopup="true" aria-controls="c-menu-panel-firefox">{{ ftl('navigation-v2-firefox-browsers') }}</a>
+  <a class="c-menu-title" href="{{ url('firefox') }}" aria-haspopup="true" aria-controls="c-menu-panel-firefox">{{ ftl('navigation-v2-firefox-browsers') }}</a>
   <div class="c-menu-panel" id="c-menu-panel-firefox">
     <div class="c-menu-panel-container">
       <button class="c-menu-button-close" type="button" aria-controls="c-menu-panel-firefox">{{ ftl('navigation-v2-close-firefox-browsers-menu') }}</button>
@@ -91,7 +91,7 @@
             </section>
           </li>
         </ul>
-        <p class="c-menu-category-link"><a href="{{ url('firefox.browsers.index') }}" data-link-text="View all Firefox Browsers" data-link-type="nav" data-link-position="topnav" data-link-group="firefox">{{ ftl('navigation-v2-view-all-firefox-browsers') }}</a></p>
+        <p class="c-menu-category-link"><a href="{{ url('firefox') }}" data-link-text="View all Firefox Browsers" data-link-type="nav" data-link-position="topnav" data-link-group="firefox">{{ ftl('navigation-v2-view-all-firefox-browsers') }}</a></p>
       </div>
     </div><!-- close .c-menu-panel-container -->
   </div><!-- close .c-menu-panel -->
