+++ bedrock/contentful/templates/includes/contentful/cta.html
@@ -16,7 +16,7 @@
 
 {% elif action == 'Explore Firefox' %}
 
-  <a href="{{ url('firefox.browsers.index') }}" class="mzp-c-button {{ button_class }}" data-cta-text="{{ cta_text }}" data-cta-type="link">{{ label }}</a>
+  <a href="{{ url('firefox') }}" class="mzp-c-button {{ button_class }}" data-cta-text="{{ cta_text }}" data-cta-type="link">{{ label }}</a>
 
 {% elif action == 'Create a Firefox Account' %}
