+++ bedrock/firefox/templates/firefox/browsers/what-is-a-browser.html
@@ -21,7 +21,7 @@
       <div class="mzp-l-content">
         <div class="c-sub-navigation-content">
           <h2 class="c-sub-navigation-title is-summary">
-            <a href="{{ url('firefox.browsers.index') }}" data-link-type="nav" data-link-position="subnav" data-link-text="Firefox">
+            <a href="{{ url('firefox') }}" data-link-type="nav" data-link-position="subnav" data-link-text="Firefox">
               <img class="c-sub-navigation-icon" src="{{ static('protocol/img/logos/firefox/logo.svg') }}" width="24" height="24" alt="">
               {{ ftl('sub-navigation-firefox') }}
             </a>
