+++ bedrock/firefox/templates/firefox/challenge-the-default/landing-switch.html
@@ -17,5 +17,5 @@ file, You can obtain one at https://mozilla.org/MPL/2.0/.
     {% include 'firefox/challenge-the-default/langs/landing.pl.html' %}
   {% endif %}
 {% else %}
-  {% include 'firefox/home/index-master.html' %}
+  {% include 'firefox/index.html' %}
 {% endif %}
