+++ bedrock/firefox/templates/firefox/channel/base.html
@@ -24,7 +24,7 @@
 {{ sub_nav(
   title={
     'text':  ftl('sub-navigation-firefox'),
-    'href':  url('firefox.browsers.index'),
+    'href':  url('firefox'),
     'cta_name': "Firefox",
     'icon': static('protocol/img/logos/firefox/logo.svg')
   },
