+++ bedrock/firefox/templates/firefox/faq.html
@@ -64,7 +64,7 @@
       <li>
         <h2>{{ ftl('what-is-firefox') }}</h2>
         <p>{{ ftl('the-firefox-browser',
-              url=url('firefox.browsers.index'),
+              url=url('firefox'),
               url2=url('products.landing')) }}</p>
       </li>
       <li>
