+++ bedrock/firefox/templates/firefox/index.html
@@ -6,7 +6,6 @@
 
 {% extends "firefox/base/base-protocol.html" %}
 {% from "macros.html" import google_play_button, apple_app_store_button, fxa_email_form with context %}
-{% from "macros-protocol.html" import split with context %}
 
 {# Issue 13019: Avoid duplicate content for English pages. #}
 {%- block page_title_full -%}
@@ -24,7 +23,6 @@
 {% set show_firefox_app_store_banner = switch('firefox-app-store-banner') %}
 
 {% block page_css %}
-  {{ css_bundle('protocol-split') }}
   {{ css_bundle('protocol-emphasis-box')}}
   {{ css_bundle('firefox-browsers-products') }}
 
@@ -39,10 +37,10 @@
   {% endif %}
 {% endblock %}
 
-{% set android_url = play_store_url('firefox', 'firefox-browsers') %}
-{% set ios_url = app_store_url('firefox', 'firefox-browsers') %}
-{% set referrals = '?utm_source=www.mozilla.org&utm_medium=referral&utm_campaign=firefox-browsers' %}
-{% set _entrypoint = 'mozilla.org-firefox-browsers' %}
+{% set android_url = play_store_url('firefox', 'firefox-home') %}
+{% set ios_url = app_store_url('firefox', 'firefox-home') %}
+{% set referrals = '?utm_source=www.mozilla.org&utm_medium=referral&utm_campaign=firefox-home' %}
+{% set _entrypoint = 'mozilla.org-firefox-home' %}
 
 {% block sub_navigation %}
   {% if ftl_has_messages('sub-navigation-firefox', 'sub-navigation-what-is-a-browser', 'sub-navigation-browsing-history') %}
@@ -57,9 +55,25 @@
           </h2>
           <ul class="c-sub-navigation-list is-details is-closed">
             <li class="c-sub-navigation-item">
-              <a href="{{ url('firefox.browsers.what-is-a-browser') }}" data-link-type="nav" data-link-position="subnav" data-link-text="What Is a Browser?">{{ ftl('sub-navigation-what-is-a-browser') }}</a>
+              <a href="{{ url('firefox.browsers.what-is-a-browser') }}" data-link-type="nav" data-link-position="subnav" data-link-text="What Is a Browser?">
+                {{ ftl('sub-navigation-what-is-a-browser') }}
+              </a>
             </li>
-            <li class="c-sub-navigation-item"><a href="{{ url('firefox.browsers.browser-history') }}" data-link-type="nav" data-link-position="subnav" data-link-text="Browser History">{{ ftl('sub-navigation-browser-history') }}</a></li>
+            <li class="c-sub-navigation-item">
+              <a href="{{ url('firefox.browsers.browser-history') }}" data-link-type="nav" data-link-position="subnav" data-link-text="Browser History">
+                {{ ftl('sub-navigation-browser-history') }}
+              </a>
+            </li>
+            <li class="c-sub-navigation-item">
+              <a href="{{ url('firefox.faq') }}" data-link-type="nav" data-link-position="subnav" data-link-text="FAQ">
+                {{ ftl('sub-navigation-faq') }}
+              </a>
+            </li>
+            <li class="c-sub-navigation-item">
+              <a href="{{ url('firefox.more') }}" data-link-type="nav" data-link-position="subnav" data-link-text="Learn More">
+                {{ ftl('sub-navigation-learn-more') }}
+              </a>
+            </li>
           </ul>
         </div>
       </div>
@@ -69,31 +83,19 @@
 
 {% block content %}
 <main>
-  {% call split(
-    image=resp_img('img/firefox/browsers/hero-700.jpg',
-      srcset={
-        'img/firefox/browsers/hero-500.jpg': '500w',
-        'img/firefox/browsers/hero-700.jpg': '700w',
-        'img/firefox/browsers/hero-900.jpg': '900w',
-        'img/firefox/browsers/hero-1100.jpg': '1100w',
-        'img/firefox/browsers/hero-1300.jpg': '1300w',
-      },
-      sizes={
-        '(min-width: 768px)': 'calc(50vw - 192px)',
-        'default': 'calc(100vw - 48px)'
-      },
-      optional_attributes={
-        'class': 'mzp-c-split-media-asset',
-        'height': '420',
-        'width': '680'
-      }
-    ),
-    block_class='c-landing-banner mzp-l-split-center-on-sm-md',
-    media_class='mzp-l-split-media-overflow',
-    media_after=True
-  ) %}
-    <h1 class="mzp-has-zap-11">{{ ftl('firefox-browsers-get-the-browsers-strong-v2', fallback='firefox-browsers-get-the-browsers-strong') }}</h1>
-  {% endcall %}
+  <header class="c-main-header mzp-l-content mzp-t-content-md">
+    <div class="mzp-c-logo mzp-t-logo-lg mzp-t-product-firefox"></div>
+    {% if ftl_has_messages('firefox-browsers-main-heading') %}
+      <h1>{{ ftl('firefox-browsers-main-heading') }}</h1>
+      <h2 class="mzp-has-zap-11">
+        {{ ftl('firefox-browsers-get-the-browsers-strong-v2', fallback='firefox-browsers-get-the-browsers-strong') }}
+      </h2>
+    {% else %}
+      <h1 class="mzp-has-zap-11">
+        {{ ftl('firefox-browsers-get-the-browsers-strong-v2', fallback='firefox-browsers-get-the-browsers-strong') }}
+      </h1>
+    {% endif %}
+  </header>
 
   <div class="mzp-l-content c-landing-grid">
     <div class="c-landing-grid-item">
