+++ bedrock/firefox/urls.py
@@ -36,8 +36,6 @@ urlpatterns = (
         ),
     ),
     path("firefox/all/", views.firefox_all, name="firefox.all"),
-    page("firefox/browsers/", "firefox/browsers/index.html", ftl_files=["firefox/browsers"]),
-    page("firefox/products/", "firefox/products/index.html", ftl_files=["firefox/products"]),
     page("firefox/channel/desktop/", "firefox/channel/desktop.html", ftl_files=["firefox/channel"]),
     page("firefox/channel/android/", "firefox/channel/android.html", ftl_files=["firefox/channel"]),
     page("firefox/channel/ios/", "firefox/channel/ios.html", ftl_files=["firefox/channel"]),
