+++ bedrock/mozorg/templates/mozorg/account.html
@@ -69,7 +69,7 @@
       <div class="accounts-products">
         <h2>{{ ftl('mozilla-account-sign-in-to', fallback='firefox-accounts-sign-in-to') }}</h2>
 
-        <a href="{{ url('firefox.browsers.index') }}"><h4 class="mzp-c-wordmark mzp-t-wordmark-sm mzp-t-product-firefox">{{ ftl('firefox-accounts-firefox-browser') }}</h4></a>
+        <a href="{{ url('firefox') }}"><h4 class="mzp-c-wordmark mzp-t-wordmark-sm mzp-t-product-firefox">{{ ftl('firefox-accounts-firefox-browser') }}</h4></a>
         <ul class="mzp-u-list-styled">
           {% if ftl_has_messages('firefox-accounts-send-a-tab', 'firefox-accounts-save-and-sync') %}
