+++ bedrock/newsletter/templates/newsletter/knowledge-is-power.html
@@ -32,7 +32,7 @@
           </li>
           {% if not isRise25 %}
             <li>
-              <a href="{{ url('firefox.browsers.index') }}" class="mzp-c-logo mzp-t-logo-md mzp-t-product-firefox">{{ ftl('knowledge-is-power-firefox') }}</a>
+              <a href="{{ url('firefox') }}" class="mzp-c-logo mzp-t-logo-md mzp-t-product-firefox">{{ ftl('knowledge-is-power-firefox') }}</a>
             </li>
             <li>
               <a href="{{ url('products.vpn.landing') }}" class="mzp-c-logo mzp-t-logo-md mzp-t-product-vpn">{{ ftl('knowledge-is-power-vpn') }}</a>
