+++ bedrock/products/templates/products/landing.html
@@ -71,7 +71,7 @@
           'height': '247'
         }
       ) }}
-      <h2 class="c-landing-grid-title "><a href="{{ url('firefox.browsers.index') }}" data-cta-type="link" data-cta-text="Firefox browsers">{{ ftl('firefox-products-firefox', fallback = 'firefox-products-firefox-browsers') }}</a></h2>
+      <h2 class="c-landing-grid-title "><a href="{{ url('firefox') }}" data-cta-type="link" data-cta-text="Firefox browsers">{{ ftl('firefox-products-firefox', fallback = 'firefox-products-firefox-browsers') }}</a></h2>
       <p>{{ ftl('firefox-products-get-the-browsers-that-block') }}</p>
       <div id="menu-browsers-wrapper" class="mzp-c-menu-list mzp-t-cta mzp-t-download">
         <h3 class="mzp-c-menu-list-title">{{ ftl('download-button-download') }}</h3>
@@ -88,7 +88,7 @@
           <li class="mzp-c-menu-list-item menu-ios"><a href="{{ ios_url }}" rel="external noopener" class="ga-product-download" data-link-type="download" data-display-name="iOS" data-download-version="ios" data-download-os="iOS">{{ ftl('firefox-products-ios') }}</a></li>
         </ul>
       </div>
-      <p><a class="mzp-c-cta-link" href="{{ url('firefox.browsers.index') }}" data-cta-type="link" data-cta-text="See all browsers">{{ ftl('firefox-products-see-all-browsers') }}</a></p>
+      <p><a class="mzp-c-cta-link" href="{{ url('firefox') }}" data-cta-type="link" data-cta-text="See all browsers">{{ ftl('firefox-products-see-all-browsers') }}</a></p>
     </div>
 
   {% if ftl_has_messages('firefox-products-surf-stream-and-get-work-done') %}
