--- l10n/en/firefox/home.ftl
@@ -1,67 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-### URL: https://www-dev.allizom.org/firefox/
-
-firefox-home-firefox-protect-your = { -brand-name-firefox } - Protect your life online with privacy-first products
-firefox-home-firefox-is-more-than = { -brand-name-firefox } is more than a browser. Learn more about { -brand-name-firefox } products that handle your data with respect and are built for privacy anywhere you go online.
-
-# Obsolete string
-firefox-home-the-browser-is-just = The browser is just the beginning
-
-firefox-home-firefox-is-just = { -brand-name-firefox } is just the beginning.
-
-# Obsolete string
-firefox-home-meet-our-family-of = Meet our family of products
-
-firefox-home-meet-mozillas-family  = Meet { -brand-name-mozilla }’s family of products
-firefox-home-get-trackers-off = Get 2,000+ trackers off your trail — including { -brand-name-facebook }
-firefox-home-know-when-hackers-strike = Know when hackers strike — and stay a step ahead
-firefox-home-start-getting-breach = Start getting breach reports
-firefox-home-un-spam-your-life = Un-spam your life with { -brand-name-relay }
-firefox-home-try-relay = Try { -brand-name-relay }
-
-# The strong tags around "respect" add a special underline. The underline breaks if it is on two words, so please omit the strong tags if they need to be around multiple words in your language.
-firefox-home-get-the-respect-you = Get the <strong>respect</strong> you deserve
-
-# Obsolete string
-firefox-home-every-single-firefox = Every single { -brand-name-firefox } product honors our Personal Data Promise: <strong>Take less. Keep it safe. No secrets.</strong>
-
-firefox-home-every-single-mozilla = Every single { -brand-name-mozilla } product honors our Personal Data Promise: <strong>Take less. Keep it safe. No secrets.</strong>
-
-firefox-home-trade-clickbait-for = Trade clickbait for quality content
-firefox-home-learn-more-about-pocket = Learn more about { -brand-name-pocket }
-firefox-home-protection-for-your-whole = Protection for your whole device, on every device.
-
-# The strong tags around "privacy" add a special underline. The underline breaks if it is on two words, so please omit the strong tags if they need to be around multiple words in your language.
-firefox-home-one-login-all-your = One login. All your devices. A family of products that respect your <strong>privacy</strong>.
-
-firefox-home-get-the-browser-extension = Get the browser extension
-firefox-home-get-the-facebook-container = Get the { -brand-name-facebook-container } extension
-firefox-home-download-the-browser = Download the browser
-firefox-home-download-the-app = Download the app
-firefox-home-get-mozilla-vpn = Get { -brand-name-mozilla-vpn }
-firefox-home-desktop = Desktop
-firefox-home-browsers = Browsers
-firefox-home-android = { -brand-name-android }
-firefox-home-ios = { -brand-name-ios }
-firefox-home-monitor = { -brand-name-monitor }
-firefox-home-lockwise = { -brand-name-lockwise }
-firefox-home-send = { -brand-name-send }
-firefox-home-mozilla = { -brand-name-mozilla }
-firefox-home-pocket = { -brand-name-pocket }
-firefox-home-mozilla-vpn = { -brand-name-mozilla-vpn }
-firefox-home-relay = { -brand-name-relay }
-
-## The strings below are visually hidden in the page and replaced by logo wordmark images. They are still important for a11y and SEO.
-
-firefox-home-firefox-browser = { -brand-name-firefox-browser }
-firefox-home-mozilla-monitor = { -brand-name-mozilla-monitor }
-
-# Obsolete string
-firefox-home-firefox-monitor = { -brand-name-firefox-monitor }
-
-firefox-home-firefox-lockwise = { -brand-name-firefox-lockwise }
-firefox-home-firefox-send = { -brand-name-firefox-send }
-firefox-home-firefox-relay = { -brand-name-firefox-relay }
