+++ media/css/firefox/browsers-products.scss
@@ -14,7 +14,7 @@ $brand-theme: 'firefox';
 
 // Hello developer in the future! This file serves multiple different pages
 // which isn't so obvious, so make changes with care! Be sure to check the
-// following pages: /firefox/browsers/, /firefox/mobile/.
+// following pages: /firefox/, /firefox/mobile/.
 // See issue https://github.com/mozilla/bedrock/issues/11405
 
 $send-svg: '<svg width="24px" height="24px" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg"><g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd"><g transform="translate(3.000000, 1.000000)" stroke="#{$color-link-hover}" stroke-width="2"><path d="M2,16 L2,20 C2,21.1045695 2.8954305,22 4,22 L14,22 C15.1045695,22 16,21.1045695 16,20 L16,2 C16,0.8954305 15.1045695,-2.02906125e-16 14,0 L4,0 C2.8954305,2.02906125e-16 2,0.8954305 2,2 L2,6" stroke-linecap="round"></path><path d="M2,4 L16,4"></path><path d="M2,18 L16,18"></path><path d="M3.55271368e-15,11 L10,11" stroke-linecap="round" stroke-linejoin="round"></path><polyline stroke-linecap="round" stroke-linejoin="round" points="6 7 10 11 6 15"></polyline></g></g></svg>';
@@ -40,6 +40,14 @@ $h-grid-lg: $layout-xl;
 
 .hidden {
     display: none !important; /* stylelint-disable-line declaration-no-important */
+}
+
+.c-main-header {
+    text-align: center;
+
+    .mzp-c-logo {
+        margin: 0 auto $spacing-lg;
+    }
 }
 
 // * -------------------------------------------------------------------------- */
