--- media/css/firefox/home/ie9.scss
@@ -1,48 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-$font-path: '/media/protocol/fonts';
-$image-path: '/media/protocol/img';
-
-@import '~@mozilla-protocol/core/protocol/css/includes/lib';
-
-.js-fx-only {
-    display: none;
-}
-
-.c-showcase {
-    display: table;
-    margin-bottom: $layout-md;
-    width: 100%;
-
-    .mzp-l-content {
-        width: 100%;
-        display: table;
-    }
-}
-
-.c-showcase-title {
-    display: table-caption;
-    text-align: center;
-}
-
-.c-showcase-list {
-    display: table-row;
-
-    li {
-        display: table-cell;
-        text-align: center;
-    }
-}
-
-// issue 13317
-#menu-browsers {
-    .menu-desktop {
-        display: none;
-    }
-
-    .menu-desktop-unsupported {
-        display: block;
-    }
-}
