--- media/css/firefox/home/master.scss
@@ -1,236 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-$font-path: '/media/protocol/fonts';
-$image-path: '/media/protocol/img';
-$brand-theme: 'firefox';
-
-@import '~@mozilla-protocol/core/protocol/css/includes/lib';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/wordmark';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/wordmark-product-firefox';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/wordmark-product-monitor';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/wordmark-product-vpn';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/wordmark-product-pocket';
-@import '~@mozilla-protocol/core/protocol/css/components/logos/wordmark-product-relay';
-@import '~@mozilla-protocol/core/protocol/css/components/zap';
-@import '../../protocol/components/custom-menu-list';
-
-// conditional content
-.js-fx-only {
-    display: none;
-
-    .is-firefox & {
-        display: block;
-    }
-}
-
-#menu-browsers {
-    .ios & .menu-desktop,
-    .ios & .menu-android,
-    .android & .menu-desktop,
-    .android & .menu-ios {
-        a {
-            pointer-events: none;
-            color: $color-marketing-gray-50;
-        }
-    }
-
-    // issue 13317
-    .menu-desktop-unsupported {
-        display: none;
-    }
-
-    .fx-unsupported & .menu-desktop {
-        display: none;
-    }
-
-    .fx-unsupported & .menu-desktop-unsupported {
-        display: block;
-    }
-}
-
-// ------------------------------------------------------------------------------
-// component over rides (content menu list, feature card)
-
-.firefox-home {
-    // menu list
-    .mzp-c-menu-list {
-        margin-bottom: $spacing-lg;
-    }
-
-    // splits
-    .mzp-c-split {
-        h4 {
-            @include text-title-sm;
-            margin-bottom: $layout-sm;
-        }
-
-        .mzp-c-split-media-asset {
-            margin: 0 auto;
-        }
-    }
-}
-
-// ------------------------------------------------------------------------------
-// Lead text and image
-
-.c-lead {
-    .mzp-l-content {
-        max-width: $content-max;
-    }
-}
-
-.c-lead-title {
-    @include font-size(40px);
-    margin-bottom: 0;
-    text-align: center;
-}
-
-@media #{$mq-md} {
-    .c-lead {
-        @include bidi (((background-position, center left 50vw, center right 50vw),));
-        background-image: url('/media/img/firefox/home/master/hero.jpg');
-        background-size: auto 300px;
-        background-repeat: no-repeat;
-    }
-
-    .c-lead-title {
-        @include bidi (((text-align, left, right),));
-        @include flexbox;
-        @include align-items(center);
-        @include font-size(48px);
-        max-width: calc(50% - #{$layout-sm});
-        min-height: 300px - $layout-xl;
-    }
-}
-
-@media #{$mq-lg} {
-    .c-lead {
-        @include at2x('/media/img/firefox/home/master/hero.jpg', contain);
-        margin-top: $spacing-lg;
-        margin-bottom: $layout-md;
-    }
-
-    .c-lead-title {
-        @include font-size(64px);
-        margin-top: $spacing-lg;
-        max-width: calc(50% - #{$layout-md});
-    }
-}
-
-// ------------------------------------------------------------------------------
-// showcase
-
-.c-showcase {
-    text-align: center;
-}
-
-.c-showcase-title {
-    @include text-title-sm;
-    margin-bottom: $spacing-lg;
-}
-
-.c-showcase-list {
-    @include flexbox;
-    flex-wrap: wrap;
-    justify-content: center;
-    margin: 0 auto ($layout-md * -1) auto; // offset the bottom magin on the individual icons
-    max-width: $content-md;
-
-    .t-small & {
-        max-width: $content-sm;
-    }
-
-    li {
-        flex-grow: 1;
-        flex-basis: 0;
-        margin: 0 $layout-sm $layout-sm;
-
-        .t-small & {
-            display: inline-block;
-            flex-basis: auto;
-            height: 40px;
-            width: 40px;
-            overflow: hidden;
-        }
-    }
-}
-
-@media #{$mq-md} {
-    .c-showcase-list {
-        margin-bottom: ($layout-lg * -1); // offset the magin on the individual icons
-
-        li {
-            margin: 0 $layout-sm $layout-lg;
-
-            .t-small & {
-                margin: 0 $spacing-md $layout-lg;
-            }
-        }
-    }
-}
-
-// ------------------------------------------------------------------------------
-// promise background colour and image
-
-.t-promise {
-    padding: 0 $layout-sm;
-
-    .mzp-c-split-bg {
-        @include light-links;
-        background-image: linear-gradient(90deg, $color-violet-80, $color-violet-90);
-        border-radius: $border-radius-md;
-        box-shadow: $box-shadow-md;
-        margin: $layout-md auto;
-
-        h3,
-        p {
-            color: $color-white;
-        }
-
-        h3 {
-            @include text-title-md;
-        }
-    }
-
-    @media #{$mq-md} {
-        margin: $layout-lg 0;
-        padding: 0 $layout-md;
-
-        .mzp-c-split-bg {
-            background-image: url('/media/img/firefox/home/master/respect.jpg'), linear-gradient(90deg, #5731ab 50%, $color-violet-90);
-            background-position: left -30px center, top left;
-            background-repeat: no-repeat;
-            background-size: 50%, auto;
-
-            @media #{$mq-high-res} {
-                background-image: url('/media/img/firefox/home/master/respect-high-res.jpg'), linear-gradient(90deg, #5731ab 50%, $color-violet-90);
-            }
-
-            [dir='rtl'] & {
-                background-image: url('/media/img/firefox/home/master/respect-rtl.jpg'), linear-gradient(-90deg, #5731ab 50%, $color-violet-90);
-                background-position: right -30px center, top right;
-
-                @media #{$mq-high-res} {
-                    background-image: url('/media/img/firefox/home/master/respect-rtl-high-res.jpg'), linear-gradient(-90deg, #5731ab 50%, $color-violet-90);
-                }
-            }
-        }
-
-        .mzp-c-split-body {
-            @include flexbox;
-            align-items: center;
-            min-height: 280px;
-        }
-    }
-
-    @media #{$mq-xl} {
-        padding: 0 $layout-lg;
-
-        .mzp-c-split-container {
-            background-position: left center, top left;
-            background-size: 547px auto, auto;
-        }
-    }
-}
