+++ media/static-bundles.json
@@ -602,19 +602,6 @@
     },
     {
       "files": [
-        "css/firefox/home/ie9.scss"
-      ],
-      "name": "firefox-home-ie9"
-    },
-    {
-      "files": [
-        "css/mozorg/mozilla-account-promo.scss",
-        "css/firefox/home/master.scss"
-      ],
-      "name": "firefox-master"
-    },
-    {
-      "files": [
         "css/mozorg/mozilla-account-promo.scss",
         "css/firefox/browsers-products.scss"
       ],
@@ -1455,12 +1442,6 @@
         "js/firefox/set-as-default/thanks.js"
       ],
       "name": "firefox-default-thanks"
-    },
-    {
-      "files": [
-        "js/firefox/home/master.js"
-      ],
-      "name": "firefox-master"
     },
     {
       "files": [
