+++ tests/functional/firefox/test_home.py
@@ -2,6 +2,8 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+from urllib.parse import unquote
+
 import pytest
 
 from pages.firefox.home import FirefoxHomePage
@@ -9,7 +11,15 @@ from pages.firefox.home import FirefoxHomePage
 
 @pytest.mark.smoke
 @pytest.mark.nondestructive
-def test_download_menu_list_displays(base_url, selenium):
+def test_primary_download_button_displayed(base_url, selenium):
     page = FirefoxHomePage(selenium, base_url).open()
-    page.browser_menu_list.click()
-    assert page.browser_menu_list.list_is_open
+    assert page.is_primary_download_button_displayed
+
+
+@pytest.mark.nondestructive
+def test_account_form(base_url, selenium):
+    page = FirefoxHomePage(selenium, base_url).open()
+    page.join_firefox_form.type_email("success@example.com")
+    page.join_firefox_form.click_continue()
+    url = unquote(selenium.current_url)
+    assert "email=success@example.com" in url, "Email address is not in URL"
