+++ tests/pages/firefox/home.py
@@ -5,23 +5,18 @@
 from selenium.webdriver.common.by import By
 
 from pages.base import BasePage
-from pages.regions.menu_list import MenuList
+from pages.regions.join_firefox_form import JoinFirefoxForm
 
 
 class FirefoxHomePage(BasePage):
     _URL_TEMPLATE = "/{locale}/firefox/"
 
-    # browser download menu list
-    _browser_menu_list_locator = (By.ID, "test-menu-browsers-wrapper")
-
-    # facebook container extension link visibility
-    _facebook_container_link_locator = (By.ID, "test-fbc")
+    _primary_download_button_locator = (By.ID, "qa-desktop-download")
 
     @property
-    def fb_container_is_displayed(self):
-        return self.is_element_displayed(*self._facebook_container_link_locator)
+    def is_primary_download_button_displayed(self):
+        return self.is_element_displayed(*self._primary_download_button_locator)
 
     @property
-    def browser_menu_list(self):
-        el = self.find_element(*self._browser_menu_list_locator)
-        return MenuList(self, root=el)
+    def join_firefox_form(self):
+        return JoinFirefoxForm(self)
